amd_energy: Add family support for 64 bit RAPL registers
Some of the models in family 0x19, 0x1A support 64 bit RAPL registers.
On these platforms, amd_energy driver can be used
with some modification to read the energy.

Since X86_MODEL_ANY value is defined as 0
if we add model 0x0 in x86_cpu_id tables, x86_match_cpu()
thinks it as any model will be allowed and it does return the
descriptor for that entry.
This is not an accepted behaviour. Hence add a check for 0x0 models
separately.